Bluetooth: Cleanup BREDR profiles before stopping
Root cause: Profiles while stopping as part of transition
from TURNING_OFF to BLE_ON does cleanup, assuming the cleanup
is already done in the stack for the services. But as per the current
implementation all stack level cleanup takes place during BLE_TURNING_OFF
to OFF transition.

Fix: We are posting a property from AdapterProperties while transitioning
from TURNING_OFF to BLE_ON which leads to cleanup of BREDR profiles
(from the btif layer). A callback is initiated from btif layer which
redirects it to the original implementation of stopping the BREDR profiles.
Also, since the cleanup for BREDR profiles in the btif layer already
happened during transition to BLE_ON , the cleanup during BLE_TURNING_OFF
to OFF is limited to BLE profiles.

CRs-Fixed: 984981
Change-Id: Id1f6417945da0495cf3abd004c61464ebdeb5f50